refactor/inprovements
