+++ extension-client/src/webview/ReactView.ts
@@ -35,6 +35,7 @@ export abstract class ReactView extends View {
     }
 
     async getHTMLString(webview: vscode.Webview): Promise<string> {
+        
         const manifestPath = path.join(
             this._extensionPath,
             'dist',
